Adds an exception_notifier callback support for endpoints
The Connection, Session, Receiver, and Sender endpoints now
support a callback which is called when an exception occurs.
The handler of the callback is register-able through a public
method. The callback is optional and will only be called on
an error if a handler is set.

See the method docstrings for more information.